moved MenuPicker to common folder
